# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Workshop Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/moderation-actions": {
            "get": {
                "security": [
                    {
                        "DebugUserRoles": []
                    },
                    {
                        "DebugUserID": []
                    }
                ],
                "description": "Returns a list of moderation actions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Moderation"
                ],
                "summary": "Get Moderation Actions",
                "parameters": [
                    {
                        "enum": [
                            "approve",
                            "decline"
                        ],
                        "type": "string",
                        "description": "Moderation action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order by creation time",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of moderation actions per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of moderation actions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/workshop.ModerationAction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request – invalid payload",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden – caller lacks post moderator permissions",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
...
        "/posts/{postID}/favorite": {
            "post": {
                "security": [
                    {
                        "DebugUserRoles": []
                    },
                    {
                        "DebugUserID": []
                    }
                ],
                "description": "Mark a post as favorite for the current user. Only authorized users with role 'USER' can perform this action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites",
                    "Interactions"
                ],
                "summary": "Favorite Post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success – post was successfully added to favorites (no content returned)"
                    },
                    "400": {
                        "description": "Bad Request – invalid input payload",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden – only authorized users can favorite posts",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found – post not found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict – post is already in favorites",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "DebugUserRoles": []
                    },
                    {
                        "DebugUserID": []
                    }
                ],
                "description": "Removes a post from the current user's favorites. Only authorized users with role 'USER' can perform this action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites",
                    "Interactions"
                ],
                "summary": "Unfavorite Post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content – the post was successfully removed from favorites"
                    },
                    "400": {
                        "description": "Bad Request – invalid input payload",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden – only authorized users can unfavorite posts",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found – post is not marked as favorite or does not exist\"\t//\t(рекомендуемый статус: 412 Precondition Failed, если статус отличается)",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
        "/posts/{postID}/moderate": {
            "post": {
                "security": [
                    {
                        "DebugUserRoles": []
                    },
                    {
                        "DebugUserID": []
                    }
                ],
                "description": "Allows a user with the 'POST_MODERATOR' role to moderate a post.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts",
                    "Moderation"
                ],
                "summary": "Moderate Post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request payload containing moderation action and post identifier",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ModeratePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success – the moderation action has been applied (no content returned)"
                    },
                    "400": {
                        "description": "Bad Request – invalid payload",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden – caller lacks post moderator permissions",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
        "/posts/{postID}/rate": {
            "post": {
                "security": [
                    {
                        "DebugUserRoles": []
                    },
                    {
                        "DebugUserID": []
                    }
                ],
                "description": "Allows an authorized user (role \"user\") to rate a post.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Posts",
                    "Interactions"
                ],
                "summary": "Rate Post",
                "parameters": [
                    {
                        "description": "Request payload containing the post identifier and rating data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.RatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success – the post rating has been processed successfully (no content returned)"
                    },
                    "400": {
                        "description": "Bad Request – invalid input or payload",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden – only authorized users can rate posts",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found – the post was not found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict – the post has already been rated by the user",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controller.ModeratePostRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "enum": [
                        "approve",
                        "decline"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ModerationActionType"
                        }
                    ],
                    "x-order": "0"
                },
                "note": {
                    "type": "string",
                    "x-order": "1"
                }
            }
        },
        "controller.RateAction": {
            "type": "string",
            "enum": [
                "upvote",
                "downvote",
                "retract"
            ],
            "x-enum-varnames": [
                "RateActionUpvote",
                "RateActionDownvote",
                "RateActionRetract"
            ]
        },
        "controller.RatePostRequest": {
            "type": "object",
            "required": [
                "rating"
            ],
            "properties": {
                "rating": {
                    "enum": [
                        "upvote",
                        "downvote",
                        "retract"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/controller.RateAction"
                        }
                    ]
                }
            }
        },
...
        "types.ModerationActionType": {
            "type": "string",
            "enum": [
                "approve",
                "decline"
            ],
            "x-enum-varnames": [
                "ModerationActionTypeApprove",
                "ModeratorActionTypeDecline"
            ]
        },
        "types.ModerationStatus": {
            "type": "string",
            "enum": [
                "approved",
                "declined",
                "pending"
            ],
            "x-enum-varnames": [
                "ModerationStatusApproved",
                "ModerationStatusDeclined",
                "ModerationStatusPending"
            ]
        },
        "types.RateType": {
            "type": "string",
            "enum": [
                "upvoted",
                "downvoted",
                "voted",
                "none"
            ],
            "x-enum-varnames": [
                "RateTypeUpvoted",
                "RateTypeDownvoted",
                "RateTypeVoted",
                "RateTypeNone"
            ]
        },
...
        "workshop.ModerationAction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "0"
                },
                "post": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/workshop.Post"
                        }
                    ],
                    "x-order": "1"
                },
                "moderator_id": {
                    "type": "string",
                    "x-order": "2"
                },
                "action": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ModerationActionType"
                        }
                    ],
                    "x-order": "3"
                },
                "note": {
                    "type": "string",
                    "x-order": "4"
                },
                "created_at": {
                    "type": "string",
                    "x-order": "5"
                }
            }
        },
        "workshop.Post": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "0"
                },
                "title": {
                    "type": "string",
                    "x-order": "1"
                },
                "interaction_data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/workshop.PostInteractionData"
                        }
                    ],
                    "x-order": "10"
                },
                "rating": {
                    "type": "integer",
                    "x-order": "11"
                },
                "comments_count": {
                    "type": "integer",
                    "x-order": "12"
                },
                "favorites_count": {
                    "type": "integer",
                    "x-order": "13"
                },
                "description": {
                    "type": "string",
                    "x-order": "2"
                },
                "preview_url": {
                    "type": "string",
                    "x-order": "3"
                },
                "post_type": {
                    "type": "string",
                    "x-order": "4"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-order": "5"
                },
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workshop.PostContent"
                    },
                    "x-order": "6"
                },
                "created_at": {
                    "type": "string",
                    "x-order": "7"
                },
                "updated_at": {
                    "type": "string",
                    "x-order": "8"
                },
                "moderation_data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/workshop.PostModerationData"
                        }
                    ],
                    "x-order": "9"
                }
            }
        },
        "workshop.PostContent": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "0"
                },
                "content_type": {
                    "type": "string",
                    "x-order": "1"
                },
                "content_data": {
                    "type": "string",
                    "x-order": "2"
                },
                "is_link": {
                    "type": "boolean",
                    "x-order": "3"
                }
            }
        },
        "workshop.PostInteractionData": {
            "type": "object",
            "properties": {
                "is_favorite": {
                    "type": "boolean",
                    "x-order": "0"
                },
                "vote": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.RateType"
                        }
                    ],
                    "x-order": "1"
                }
            }
        },
        "workshop.PostModerationData": {
            "type": "object",
            "properties": {
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ModerationStatus"
                        }
                    ],
                    "x-order": "0"
                },
                "note": {
                    "type": "string",
                    "x-order": "1"
                }
            }
        }
    },
    "securityDefinitions": {
        "DebugUserID": {
            "description": "Force user ID if debug mode is enabled. Required if Debug User Roles is provided.",
            "type": "apiKey",
            "name": "X-Debug-User-ID",
            "in": "header"
        },
        "DebugUserRoles": {
            "description": "Force user roles if debug mode is enabled. Split by comma ',' for multiple roles.",
            "type": "apiKey",
            "name": "X-Debug-User-Roles",
            "in": "header"
        }
    }
}